requirements.txt
flask==3.1.0
pillow==10.3.0
webcolors==1.13
